move to using roaring iterators for UnmarshalBinary
We have this fancy roaring iterator, let's try it out.

This isn't fully debugged, but the basic idea is that the roaring
iterator may be faster, and the overhead of PutContainerValues
plus a separate iteration may be expensive. Note that we're
not calling NewContainer now, which is arguably a design flaw,
but it should let us test the thing.